hakhathon_2/requirements.txt
Flask==2.2.3
Flask_Migrate==4.0.4
flask_sqlalchemy==3.0.3
